coralscapesScripts/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.4
albumentations>=2.0.0
segmentation_models_pytorch>=0.5.0
peft>=0.17.0
accelerate>=1.10.0
torchmetrics>=1.8.0
wandb>=0.22.0
scikit-learn>=1.0.0
tqdm>=4.60.0
Pillow>=9.0.0
PyYAML>=6.0
matplotlib>=3.5.0
huggingface_hub>=0.25.0
transformers>=4.30.0
